backend/requirements.txt
# AI 中台后端依赖
# Python 3.10+

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Async
asyncio>=3.4.3
aiohttp>=3.9.0
celery>=5.3.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
alembic>=1.13.0

# Validation & Settings
pydantic>=2.5.0
pydantic-settings>=2.1.0

# AI/ML
torch>=2.1.0
transformers>=4.36.0
sentence-transformers>=2.2.0
langchain>=0.1.0
langchain-community>=0.0.10

# Vector Database
pymilvus>=2.3.0
qdrant-client>=1.7.0

# Model Inference
vllm>=0.2.0
openai>=1.10.0

# Document Processing
pypdf>=4.0.0
python-docx>=1.1.0
openpyxl>=3.1.0
python-pptx>=0.6.23
unstructured>=0.11.0

# Security & Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.0

# Monitoring & Logging
prometheus-client>=0.19.0
structlog>=24.1.0
loguru>=0.7.0

# HTTP Client
httpx>=0.26.0

# Redis
redis>=5.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
locust>=2.20.0  # 性能测试

# Dev Tools
black>=24.1.0
ruff>=0.1.0
mypy>=1.8.0
